requirements.txt
absl-py==0.8.0
astor==0.8.0
atari-py==0.2.6
baselines==0.1.5
bleach==1.5.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
Cython==0.29.13
#DeepMind-Lab==1.0
dill==0.3.0
future==0.17.1
gast==0.2.2
glfw==1.8.3
google-pasta==0.1.7
grpcio==1.23.0
gym==0.10.5
gym-deepmindlab==0.1.2
h5py==2.9.0
html5lib==0.9999999
idna==2.8
imageio==2.5.0
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lockfile==0.12.2
Markdown==3.1.1
mpi4py==3.0.2
numpy==1.17.1
opencv-python==4.1.1.26
Pillow==6.1.0
#pkg-resources==0.0.0
progressbar2==3.43.1
protobuf==3.9.1
pycparser==2.19
pyglet==1.4.3
PyOpenGL==3.1.0
pypng==0.0.20
python-utils==2.3.0
pyzmq==18.1.0
requests==2.22.0
scipy==1.3.1
six==1.12.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
tqdm==4.35.0
urllib3==1.25.3
Werkzeug==0.15.6
wrapt==1.11.2
zmq==0.0.0
